+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- HTML elements being rendered with ids equal to reserved global variable names, such as 'global', which would result in these variables being overwritten.
 
 ## [3.66.1] - 2020-07-31